+++ lib/kazan/models/model_desc.ex
@@ -1,7 +1,7 @@
-defmodule Kazan.Codegen.Models.ModelDesc do
+defmodule Kazan.Models.ModelDesc do
   @moduledoc false
   import Kazan.Codegen.Naming, only: [model_name_to_module: 1]
-  alias Kazan.Codegen.Models.{ResourceId, PropertyDesc}
+  alias Kazan.Models.{ResourceId, PropertyDesc}
 
   defstruct [
     :id,